requirements.txt
numpy~=1.22.3
sewar~=0.4.5
opencv-python~=4.5.5.64
matplotlib~=3.5.1
Pillow~=9.0.1